more csrf work

+++ interface/super/manage_document_templates.php
@@ -19,7 +19,7 @@ if (!acl_check('admin', 'super')) {
     die(htmlspecialchars(xl('Not authorized')));
 }
 
-$form_filename = strip_escape_custom($_REQUEST['form_filename']);
+$form_filename = convert_safe_file_dir_name($_REQUEST['form_filename']);
 
 $templatedir = "$OE_SITE_DIR/documents/doctemplates";
 
@@ -27,6 +27,11 @@ $templatedir = "$OE_SITE_DIR/documents/doctemplates";
 // Thus the current browser page should remain displayed.
 //
 if (!empty($_POST['bn_download'])) {
+    //verify csrf
+    if (!verifyCsrfToken($_POST["csrf_token_form"])) {
+        die(xlt('Authentication Error'));
+    }
+
     $templatepath = "$templatedir/$form_filename";
     header('Content-Description: File Transfer');
     header('Content-Transfer-Encoding: binary');
@@ -45,6 +50,11 @@ if (!empty($_POST['bn_download'])) {
 }
 
 if (!empty($_POST['bn_delete'])) {
+    //verify csrf
+    if (!verifyCsrfToken($_POST["csrf_token_form"])) {
+        die(xlt('Authentication Error'));
+    }
+
     $templatepath = "$templatedir/$form_filename";
     if (is_file($templatepath)) {
         unlink($templatepath);
@@ -52,6 +62,11 @@ if (!empty($_POST['bn_delete'])) {
 }
 
 if (!empty($_POST['bn_upload'])) {
+    //verify csrf
+    if (!verifyCsrfToken($_POST["csrf_token_form"])) {
+        die(xlt('Authentication Error'));
+    }
+
   // Handle uploads.
     $tmp_name = $_FILES['form_file']['tmp_name'];
     if (is_uploaded_file($tmp_name) && $_FILES['form_file']['size']) {
@@ -61,9 +76,13 @@ if (!empty($_POST['bn_upload'])) {
             $form_dest_filename = $_FILES['form_file']['name'];
         }
 
-        $form_dest_filename = preg_replace("/[^a-zA-Z0-9_.]/", "_", basename($form_dest_filename));
+        $form_dest_filename = convert_safe_file_dir_name(basename($form_dest_filename));
         if ($form_dest_filename == '') {
             die(htmlspecialchars(xl('Cannot determine a destination filename')));
+        }
+        $path_parts = pathinfo($form_dest_filename);
+        if (!in_array(strtolower($path_parts['extension']), array('odt','txt'))) {
+            die(text(strtolower($path_parts['extension'])) . ' ' . xlt('filetype is not accepted'));
         }
 
         $templatepath = "$templatedir/$form_dest_filename";
@@ -101,6 +120,7 @@ if (!empty($_POST['bn_upload'])) {
 <body class="body_top">
 <form method='post' action='manage_document_templates.php' enctype='multipart/form-data'
  onsubmit='return top.restoreSession()'>
+<input type="hidden" name="csrf_token_form" value="<?php echo attr($_SESSION['csrf_token']); ?>" />
 
 <center>
